requirements.txt
click==8.0.1
EasyProcess==0.3
entrypoint2==0.2.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
patool==1.12
pyunpack==0.2.2
Werkzeug==2.0.1
